taro-api/requirements.txt
# LangGraph + LangChain
langgraph>=0.6.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
# Optional: additional LLM providers
langchain-anthropic>=0.3.0
langchain-google-genai>=2.1.0

# SurrealDB ecosystem
langchain-surrealdb
langgraph-checkpoint-surrealdb
surrealdb

# Optional: policy ingestion queue (Layer 2)
redis>=5.0.0

# Search
langchain-tavily

# API
fastapi
uvicorn[standard]
pydantic>=2.0

# Utilities
python-dotenv
loguru
langsmith

# Testing
pytest
pytest-asyncio
httpx
